fixed skip type

+++ extern-api-java-sdk-common/src/main/java/ru/kontur/extern_api/sdk/model/DocflowFilter.java
@@ -34,7 +34,7 @@ import ru.kontur.extern_api.sdk.utils.YAStringUtils;
 
 public class DocflowFilter implements Filter {
 
-    private long skip;
+    private int skip;
     private int take = 1000;
     private SortOrder orderBy;
 
@@ -44,7 +44,7 @@ public class DocflowFilter implements Filter {
 
     }
 
-    private DocflowFilter skip(long skip) {
+    private DocflowFilter skip(int skip) {
         this.skip = skip;
         return this;
     }
@@ -147,7 +147,7 @@ public class DocflowFilter implements Filter {
     }
 
     @Override
-    public long getSkip() {
+    public int getSkip() {
         return skip;
     }
 
@@ -161,7 +161,7 @@ public class DocflowFilter implements Filter {
         return Filter.stripNullKeys(filterMap);
     }
 
-    public static DocflowFilter page(long skip, int take) {
+    public static DocflowFilter page(int skip, int take) {
         return new DocflowFilter()
                 .skip(skip)
                 .take(take);
