+++ extern-api-java-sdk-common/src/main/java/ru/kontur/extern_api/sdk/model/Filter.java
@@ -29,7 +29,7 @@ import java.util.stream.Collectors;
 
 public interface Filter {
 
-    long getSkip();
+    int getSkip();
 
     int getTake();
