+++ extern-api-java-sdk/src/main/java/ru/kontur/extern_api/sdk/service/impl/OrganizationServiceImpl.java
@@ -128,7 +128,7 @@ public class OrganizationServiceImpl implements OrganizationService {
     public CompletableFuture<QueryContext<OrganizationBatch>> searchAsync(
             @Nullable String inn,
             @Nullable String kpp,
-            @Nullable Long skip,
+            @Nullable Integer skip,
             @Nullable Integer take
     ) {
         return searchAsync(OrgFilter
